fix: must override default chunk loader function from webpack plugin
